requirements.txt
Django==4.0
djangorestframework
python-decouple
psycopg2-binary
